+++ pkg/src/proj6.cpp
@@ -1092,12 +1092,12 @@ PROJcopyEPSG(SEXP tf) {
     INTEGER_POINTER(ans)[0] = 0;
     int i, crs_cnt;
     PROJ_CRS_INFO **proj_crs_info;
-    //PJ_CONTEXT *ctx = proj_context_create();
+    PJ_CONTEXT *ctx = proj_context_create();
     FILE *fptf;
 
 	
     crs_cnt = 0;
-    proj_crs_info = proj_get_crs_info_list_from_database(PJ_DEFAULT_CTX, "EPSG", NULL,
+    proj_crs_info = proj_get_crs_info_list_from_database(ctx, "EPSG", NULL,
         &crs_cnt);
     if (crs_cnt < 1) {
         UNPROTECT(1);
@@ -1112,25 +1112,26 @@ PROJcopyEPSG(SEXP tf) {
 
 //    PJ *pj = NULL;
 // blocks error messages in this context
-    proj_log_func(PJ_DEFAULT_CTX, NULL, proj_logger);
+    proj_log_func(ctx, NULL, proj_logger);
     for (i = 0; i < crs_cnt; i++) {
         const char *proj_definition;
 
-        PJ* pj = proj_create_from_database(PJ_DEFAULT_CTX,
+        PJ* pj = proj_create_from_database(ctx,
             proj_crs_info[i]->auth_name, proj_crs_info[i]->code,
             PJ_CATEGORY_CRS, 0, NULL);
-        proj_definition = proj_as_proj_string(PJ_DEFAULT_CTX, pj,
+        proj_definition = proj_as_proj_string(ctx, pj, // valgrind
             PJ_PROJ_5, NULL);
-        proj_destroy(pj);
+//        proj_destroy(pj); // valgrind
 
-        fprintf(fptf, "%s,\"%s\",\"%s\",\"%s\"\n", proj_crs_info[i]->code,
+        fprintf(fptf, "%s,\"%s\",\"%s\",\"%s\"\n", proj_crs_info[i]->code, //ASAN
   	    proj_crs_info[i]->name, proj_definition, 
             proj_crs_info[i]->projection_method_name);
+        proj_destroy(pj);
     }
 
     fclose(fptf);
     proj_crs_info_list_destroy(proj_crs_info);
-    //proj_context_destroy(ctx);
+    proj_context_destroy(ctx);
     INTEGER_POINTER(ans)[0] = crs_cnt;
     UNPROTECT(1);